pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seep"
version = "2.0.0"
description = "Seep - Windows Privilege Escalation Enumeration Framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
seep = "server.cli:main"

[tool.setuptools.packages.find]
include = ["server*"]

[tool.setuptools.package-data]
"server.agent.checks" = ["*.ps1"]
"server.agent.templates" = ["*.ps1"]
"server.catalog" = ["*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
